I don't even know
This was a pretty poor attempt at downgrading this mod to 1.19.2. It is full of errors.
For some reason, it's like the stoneage - I have to make EVERYTHING from scratch. And I mean, everything.
Every widget must be copied and every piece of code must be translated to the older version.
I don't think it's worth my time that much. Might come back to it, might not.