Bump version to v3.0
The decision behind this is that this is 99% the last feature version, and a rounded release suits the finish touches. This was a humongous update, too.
Nevertheless, my goal is that every release of new Genshin insturments, the major version will go up with it such that it will match the supported Genshin version.